// src/app/models/countries.model.ts
export interface Country {
  name:         Name;
  tld?:         string[];
  cca2:         string;
  ccn3?:        string;
  cca3:         string;
  cioc?:        string;
  independent?: boolean;
  status:       Status;
  unMember:     boolean;
  currencies?:  Currencies;
  idd:          Idd;
  capital?:     string[];
  altSpellings: string[];
  region:       Region;
  subregion?:   string;
  languages?:   { [key: string]: string };
  translations: { [key: string]: Translation };
  latlng:       number[];
  landlocked:   boolean;
  area:         number;
  demonyms?:    Demonyms;
  flag:         string;
  maps:         Maps;
  population:   number;
  fifa?:        string;
  car:          Car;
  timezones:    string[];
  continents:   Continent[];
  flags:        CoatOfArms;
  coatOfArms:   CoatOfArms;
  startOfWeek:  StartOfWeek;
  capitalInfo:  CapitalInfo;
  borders?:     string[];
  gini?:        { [key: string]: number };
  postalCode?:  PostalCode;
}

export interface CapitalInfo {
  latlng?: number[];
}

export interface Car {
  signs?: string[];
  side:   Side;
}

export enum Side {
  Left = "left",
  Right = "right",
}

export interface CoatOfArms {
  png?: string;
  svg?: string;
}

export enum Continent {
  Africa = "Africa",
  Antarctica = "Antarctica",
  Asia = "Asia",
  Europe = "Europe",
  NorthAmerica = "North America",
  Oceania = "Oceania",
  SouthAmerica = "South America",
}

export interface Currencies {
  XCD?: TartuGecko;
  MYR?: TartuGecko;
  EUR?: TartuGecko;
  CKD?: TartuGecko;
  NZD?: TartuGecko;
  XOF?: TartuGecko;
  LYD?: TartuGecko;
  DZD?: TartuGecko;
  MAD?: TartuGecko;
  MRU?: TartuGecko;
  SBD?: TartuGecko;
  TND?: TartuGecko;
  GBP?: TartuGecko;
  GGP?: TartuGecko;
  SCR?: TartuGecko;
  USD?: TartuGecko;
  GEL?: TartuGecko;
  SYP?: TartuGecko;
  NOK?: TartuGecko;
  PHP?: TartuGecko;
  ISK?: TartuGecko;
  MZN?: TartuGecko;
  AFN?: TartuGecko;
  DJF?: TartuGecko;
  PEN?: TartuGecko;
  BIF?: TartuGecko;
  GYD?: TartuGecko;
  AUD?: TartuGecko;
  EGP?: TartuGecko;
  ILS?: TartuGecko;
  JOD?: TartuGecko;
  BBD?: TartuGecko;
  KID?: TartuGecko;
  ZAR?: TartuGecko;
  MNT?: TartuGecko;
  MUR?: TartuGecko;
  BMD?: TartuGecko;
  ANG?: TartuGecko;
  JEP?: TartuGecko;
  KHR?: TartuGecko;
  SDG?: BAM;
  GNF?: TartuGecko;
  IRR?: TartuGecko;
  SEK?: TartuGecko;
  BDT?: TartuGecko;
  SHP?: TartuGecko;
  MVR?: TartuGecko;
  XPF?: TartuGecko;
  RWF?: TartuGecko;
  RUB?: TartuGecko;
  BYN?: TartuGecko;
  GIP?: TartuGecko;
  CLP?: TartuGecko;
  AMD?: TartuGecko;
  HTG?: TartuGecko;
  CDF?: TartuGecko;
  IQD?: TartuGecko;
  DKK?: TartuGecko;
  NGN?: TartuGecko;
  MWK?: TartuGecko;
  COP?: TartuGecko;
  GTQ?: TartuGecko;
  FKP?: TartuGecko;
  TTD?: TartuGecko;
  KPW?: TartuGecko;
  TVD?: TartuGecko;
  IDR?: TartuGecko;
  ETB?: TartuGecko;
  BOB?: TartuGecko;
  SLL?: TartuGecko;
  XAF?: TartuGecko;
  UYU?: TartuGecko;
  TWD?: TartuGecko;
  BWP?: TartuGecko;
  ZMW?: TartuGecko;
  MOP?: TartuGecko;
  FJD?: TartuGecko;
  BZD?: TartuGecko;
  ZWL?: TartuGecko;
  QAR?: TartuGecko;
  AOA?: TartuGecko;
  MMK?: TartuGecko;
  OMR?: TartuGecko;
  WST?: TartuGecko;
  BAM?: BAM;
  SZL?: TartuGecko;
  PLN?: TartuGecko;
  KMF?: TartuGecko;
  PGK?: TartuGecko;
  KES?: TartuGecko;
  VND?: TartuGecko;
  VUV?: TartuGecko;
  UGX?: TartuGecko;
  UZS?: TartuGecko;
  PYG?: TartuGecko;
  CAD?: TartuGecko;
  THB?: TartuGecko;
  BHD?: TartuGecko;
  NPR?: TartuGecko;
  KZT?: TartuGecko;
  LAK?: TartuGecko;
  KYD?: TartuGecko;
  TMT?: TartuGecko;
  BRL?: TartuGecko;
  RON?: TartuGecko;
  CHF?: TartuGecko;
  STN?: TartuGecko;
  PKR?: TartuGecko;
  TRY?: TartuGecko;
  SRD?: TartuGecko;
  SGD?: TartuGecko;
  TZS?: TartuGecko;
  HKD?: TartuGecko;
  ALL?: TartuGecko;
  INR?: TartuGecko;
  NIO?: TartuGecko;
  JPY?: TartuGecko;
  ARS?: TartuGecko;
  CVE?: TartuGecko;
  MDL?: TartuGecko;
  PAB?: TartuGecko;
  TOP?: TartuGecko;
  CZK?: TartuGecko;
  MGA?: TartuGecko;
  HUF?: TartuGecko;
  DOP?: TartuGecko;
  JMD?: TartuGecko;
  SAR?: TartuGecko;
  AZN?: TartuGecko;
  IMP?: TartuGecko;
  CRC?: TartuGecko;
  CUC?: TartuGecko;
  CUP?: TartuGecko;
  NAD?: TartuGecko;
  RSD?: TartuGecko;
  KWD?: TartuGecko;
  UAH?: TartuGecko;
  BTN?: TartuGecko;
  YER?: TartuGecko;
  VES?: TartuGecko;
  BND?: TartuGecko;
  SSP?: TartuGecko;
  LBP?: TartuGecko;
  MKD?: TartuGecko;
  MXN?: TartuGecko;
  BSD?: TartuGecko;
  CNY?: TartuGecko;
  KGS?: TartuGecko;
  GMD?: TartuGecko;
  ERN?: TartuGecko;
  TJS?: TartuGecko;
  LKR?: TartuGecko;
  BGN?: TartuGecko;
  LSL?: TartuGecko;
  KRW?: TartuGecko;
  SOS?: TartuGecko;
  GHS?: TartuGecko;
  LRD?: TartuGecko;
  FOK?: TartuGecko;
  AED?: TartuGecko;
  HNL?: TartuGecko;
  AWG?: TartuGecko;
}

export interface TartuGecko {
  name:   string;
  symbol: string;
}

export interface BAM {
  name: string;
}

export interface Demonyms {
  eng:  Eng;
  fra?: Eng;
}

export interface Eng {
  f: string;
  m: string;
}

export interface Idd {
  root?:     string;
  suffixes?: string[];
}

export interface Maps {
  googleMaps:     string;
  openStreetMaps: string;
}

export interface Name {
  common:      string;
  official:    string;
  nativeName?: { [key: string]: Translation };
}

export interface Translation {
  official: string;
  common:   string;
}

export interface PostalCode {
  format: string;
  regex?: string;
}

export enum Region {
  Africa = "Africa",
  Americas = "Americas",
  Antarctic = "Antarctic",
  Asia = "Asia",
  Europe = "Europe",
  Oceania = "Oceania",
}

export enum StartOfWeek {
  Monday = "monday",
  Saturday = "saturday",
  Sunday = "sunday",
}

export enum Status {
  OfficiallyAssigned = "officially-assigned",
  UserAssigned = "user-assigned",
}